requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
filelock==3.13.1
fsspec==2023.10.0
idna==3.4
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
mutagen==1.47.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.1
openai-whisper==20231105
regex==2023.10.3
requests==2.31.0
sympy==1.12
tiktoken==0.5.1
torch==2.1.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
